requirements.txt
Django==3.1.4

django-environ

django-markdownx

django-dbbackup

psycopg2

gunicorn

django-heroku

whitenoise

boto3

django-storages